freebsd64: Fix the type of iov_len in struct iovec64.

+++ sys/compat/freebsd64/freebsd64.h
@@ -217,7 +217,7 @@ struct msqid_ds64 {
 
 struct iovec64 {
 	void 	*iov_base;
-	int	iov_len;
+	size_t	iov_len;
 };
 
 #endif /* !_COMPAT_CHERIABI_CHERIABI_H_ */
